aider: Added clarification on correct usage of `$` syntax in JSDoc examples.
# Aider chat conversation:

USER: Great. Can you update the JSDoc style guide to make this point clear?

ASSISTANT: Sure, I can add a section to the `style-guide/jsdoc.md` file to clarify the correct usage of examples in JSDoc comments. Here are the changes:

<source>markdown
style-guide/jsdoc.md
<<<<<<< HEAD
In all cases, adequate examples should be provided. Take note that the examples properly use the `_$` and `$` syntaxes.
=======
In all cases, adequate examples should be provided. Take note that the examples properly use the `_$` and `$` syntaxes. When using the `$` syntax, ensure that it is used correctly. For example, if you are demonstrating a type-level function that takes two arguments, the `$` syntax should be used twice, like so: `$<$<String.Append, 'bar'>, 'foo'>`.
>>>>>>> updated
</source>

This change adds a sentence to the JSDoc style guide to clarify the correct usage of the `$` syntax in examples.