aider: Wrap each command in an `aider` execution string before calling `exec`.
# Aider chat conversation:

USER: Can we modify the multi-aider to actually run aider? aider is a command line program that has interface e.g. 'aider --msg="Here is my prompt"'

ASSISTANT: Sure, we can modify the `scaffolding/multi-aider.ts` file to run the `aider` command line program instead of `echo`. Here is the change:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
  .option('template', {
    alias: 't',
    description: 'Script template',
    type: 'string',
    demandOption: true
  })
=======
  .option('template', {
    alias: 't',
    description: 'Script template. Default is \'aider --msg="Processing {s} and {t}"\'',
    type: 'string',
    default: 'aider --msg="Processing {s} and {t}"'
  })
>>>>>>> updated
</source>

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
--template 'echo Processing {s} and {t}'
=======
--template 'aider --msg="Processing {s} and {t}"'
>>>>>>> updated
</source>

This change will replace the `echo` command with the `aider` command in the script template. The `{s}` and `{t}` placeholders will be replaced with the source file and the corresponding spec file, respectively.

USER: InvalidEditBlock: edit failed!

scaffolding/multi-aider.ts does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in scaffolding/multi-aider.ts with nothing missing!

scaffolding/multi-aider.ts does not contain these 1 exact lines in a row:
```
--template 'echo Processing {s} and {t}'
```

ASSISTANT: I apologize for the confusion. It seems there was a misunderstanding. The line `--template 'echo Processing {s} and {t}'` does not exist in the `scaffolding/multi-aider.ts` file.

The `aider` command should be used in the `scaffolding/multi-aider.md` file, where the example usage is given. However, the `scaffolding/multi-aider.ts` file already has the `aider` command as the default command in the `template` option.

If you want to change the default command in the `scaffolding/multi-aider.ts` file, please provide the exact lines you want to replace.

USER: Sorry for the confusion. I actually mean, for each command we generate, that actually needs to be wrapped in the aider execution string before we call exec

ASSISTANT: I see, you want to wrap each command in an `aider` execution string before calling `exec`. We can modify the `scaffolding/multi-aider.ts` file to achieve this. Here is the change:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
  return command
=======
  return `aider --msg="${command}"`
>>>>>>> updated
</source>

This change will wrap each command in an `aider` execution string before it is returned. The `exec` function will then execute the `aider` command with the original command as a message argument.