aider: Updated multi-aider.ts to ignore files that match patterns in .multi-aider-ignore file.
# Aider chat conversation:

USER: Next, if there is a .multi-aider-ignore file, I want us to skip that file, even if it matches our glob syntax. multi-aider ignore file will be a newline-delimited list of globs to ignore

ASSISTANT: To implement this feature, we need to modify the `scaffolding/multi-aider.ts` file. We will use the `glob` package's `ignore` option to exclude files that match the patterns in the `.multi-aider-ignore` file.

Here are the steps:

1. Read the `.multi-aider-ignore` file and split its content by newline to get an array of ignore patterns.
2. Use the ignore patterns when calling `glob.sync`.

Here are the changes:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
const files = glob.sync(argv.pattern)
=======
const ignorePatterns = fs.existsSync('.multi-aider-ignore')
  ? fs.readFileSync('.multi-aider-ignore', 'utf-8').split('\n')
  : []
const files = glob.sync(argv.pattern, { ignore: ignorePatterns })
>>>>>>> updated
</source>

This change will make the script ignore files that match the patterns in the `.multi-aider-ignore` file. If the `.multi-aider-ignore` file does not exist, the script will behave as before.