aider: Refactored multi-aider to skip commands for files without corresponding spec files if the template uses `{t}`, emitting a warning in such cases.
# Aider chat conversation:

ASSISTANT: To implement this feature, we need to modify the `scaffolding/multi-aider.ts` file. We will add a check to see if the template uses `{t}` and if the corresponding spec file exists. If the template uses `{t}` and the spec file does not exist, we will skip the command and emit a warning.

Here are the changes:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
const commands = files.map((file) => {
  const specFile = file.replace('.ts', '.spec.ts')
  let command = argv.template
  if (fs.existsSync(specFile)) {
    command = command.replace('{s}', file).replace('{t}', specFile)
  } else {
    command = command.replace('{s}', file).replace('{t}', '')
  }
  return `aider --msg="${command}"`
})
=======
const commands = files.reduce((acc, file) => {
  const specFile = file.replace('.ts', '.spec.ts')
  let command = argv.template
  if (fs.existsSync(specFile)) {
    command = command.replace('{s}', file).replace('{t}', specFile)
    acc.push(`aider --msg="${command}"`)
  } else if (command.includes('{t}')) {
    console.warn(`Warning: Skipping command for ${file} as no corresponding spec file exists and the template uses {t}`)
  } else {
    command = command.replace('{s}', file).replace('{t}', '')
    acc.push(`aider --msg="${command}"`)
  }
  return acc
}, [])
>>>>>>> updated
</source>

This change will skip commands for files that do not have a corresponding spec file if the template uses `{t}`. A warning will be emitted in such cases. If the template does not use `{t}`, the command will be executed as before.