aider: Modified error message to indicate that the program is halting when skipping commands.
# Aider chat conversation:

USER: Oh, right now, if we skip commands we actually throw an error and stop the script. This is okay behavior, but can we modify the message to say we're halting the program?

ASSISTANT: Sure, we can modify the error message in the `generateCommand()` function in the `scaffolding/multi-aider.ts` file to indicate that the program is halting.

Here is the *edit block* for the change:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
    throw new Error(
      `Skipping command for ${file} as no corresponding spec file exists and the template uses {t}`
    )
=======
    throw new Error(
      `Skipping command for ${file} as no corresponding spec file exists and the template uses {t}. Halting program.`
    )
>>>>>>> updated
</source>