aider: Refactored `executeCommand` function to stream command results to the terminal while it's executing.
# Aider chat conversation:

USER: Can we display the streaming results of the command in the terminal while it's executing, if we're in step mode?

ASSISTANT: Yes, we can modify the `executeCommand` function in `multi-aider.ts` to stream the results of the command to the terminal while it's executing. This can be done by using the `spawn` function from the `child_process` module instead of `exec`, which allows us to handle the `stdout` and `stderr` data events and log the data as it's received.

Here is the *edit block* for the change:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
async function executeCommand(command: string): Promise<boolean> {
  const execPromisified = util.promisify(exec)
  try {
    const { stdout, stderr } = await execPromisified(
      command.split('\n  ').join(' ')
    )
    console.log(chalk.green(`stdout: ${stdout}`))
    console.error(chalk.red(`stderr: ${stderr}`))
    return true
  } catch (error) {
    console.error(chalk.red(`Error executing command: ${command}`))
    console.error(chalk.red(`exec error: ${error}`))
    return false
  }
}
=======
import { spawn } from 'child_process'

async function executeCommand(command: string): Promise<boolean> {
  return new Promise((resolve, reject) => {
    const [cmd, ...args] = command.split(' ')
    const proc = spawn(cmd, args)

    proc.stdout.on('data', (data) => {
      console.log(chalk.green(`stdout: ${data}`))
    })

    proc.stderr.on('data', (data) => {
      console.error(chalk.red(`stderr: ${data}`))
    })

    proc.on('close', (code) => {
      if (code !== 0) {
        console.error(chalk.red(`Error executing command: ${command}`))
        reject(false)
      } else {
        resolve(true)
      }
    })
  })
}
>>>>>>> updated
</source>

This change will ensure that the results of the command are streamed to the terminal while it's executing, when in step mode.