aider: Fixed command splitting to handle sentences and phrases with spaces by using a regular expression.
# Aider chat conversation:

USER: There is an issue on line 121, i.e. the command.split command. That will improperly split out the commands, since some of the commands are sentences

ASSISTANT: You're correct. The current implementation splits the command string into an array of substrings using a space as the delimiter. This can cause issues if the command contains sentences or phrases that include spaces.

To fix this, we can use a regular expression to split the command string. This regular expression will split the string on spaces, but it will ignore spaces that are within quotes. Here's the *edit block* for the change:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
const [cmd, ...args] = command.split(' ')
=======
const [cmd, ...args] = command.match(/(?:[^\s"]+|"[^"]*")+/g) || []
>>>>>>> updated
</source>

This change will ensure that the command string is split correctly, even if it contains sentences or phrases with spaces.