aider: Set the default value for the prompts to "yes".
# Aider chat conversation:

USER: Can we set the default to the prompts to be "y"?

ASSISTANT: Yes, we can set the default value for the prompts to be "yes". This can be done by setting the `default` property of the prompt object to `true`.

Here are the *edit blocks* for the changes:

For the `promptUser` function in `multi-aider.ts`:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
{
  type: 'confirm',
  name: 'proceed',
  message: `Do you want to execute the above command?`,
  default: false
}
=======
{
  type: 'confirm',
  name: 'proceed',
  message: `Do you want to execute the above command?`,
  default: true
}
>>>>>>> updated
</source>

And for the second prompt in the same function:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
{
  type: 'confirm',
  name: 'proceed',
  message: 'Do you want to proceed?',
  default: false
}
=======
{
  type: 'confirm',
  name: 'proceed',
  message: 'Do you want to proceed?',
  default: true
}
>>>>>>> updated
</source>

These changes will set the default value for the prompts to "yes".