aider: Set the output of the underlying stdout command to grey.
# Aider chat conversation:

USER: Can we set the output of the underlying stdout command to grey?

ASSISTANT: Yes, we can change the color of the output of the underlying stdout command to grey using the `chalk` library.

Here is the *edit block* for the change:

<source>typescript
scaffolding/multi-aider.ts
<<<<<<< HEAD
proc.stdout.on('data', (data) => {
  process.stdout.write(data.toString())
})
=======
proc.stdout.on('data', (data) => {
  process.stdout.write(chalk.gray(data.toString()))
})
>>>>>>> updated
</source>

This change will set the color of the stdout output to grey.