aider: Add a note in the style guide to not add JSDoc comments to unexported intermediary interfaces.
# Aider chat conversation:

USER: Is there any improvement we can make to the style guide to make it very clear to LLM agents that JSDoc shouldn't be added to intermediary unexported interfaces?

ASSISTANT: Yes, we can add a more explicit note in the style guide to make it clear that JSDoc comments should not be added to unexported intermediary interfaces. Here is the proposed change:

<source>markdown
style-guide/jsdoc.md
<<<<<<< HEAD
2-ary kinds possess an intermediate interface for currying. This interface is not exported, and thus doesn't need to be documented. Here is an example:
=======
2-ary kinds possess an intermediate interface for currying. This interface is not exported, and thus doesn't need to be documented. **Note: Do not add JSDoc comments to these unexported intermediary interfaces.** Here is an example:
>>>>>>> updated
</source>