bugfix: ease up on intermediate argument type checks
As specified in an active issue, in certain contexts the _$cast operator can be replaced with the native type-level intersection operator &. T & U & U simplifies to T & U but nested casts do not seem to collapse nearly as well. Additionally, we can combine this change with removal of argument extend checks to completely eliminate an entire line of type checker reasoning.